refactor: move home page into app folder
